outsourced methods Execute and CoerceArguments to CommanHandler class to make Program class leaner
